requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
streamlit>=1.28.0
gdown>=4.6.0